test: added fixtures
