+++ src/cli/tests/cli.spec.ts
@@ -1,7 +1,8 @@
 import * as path from 'path';
 import { exec, ExecException } from 'child_process';
 import { normalizeString } from './__utils__/normalizeString';
-
+import { asyncLoadConfig } from '../utils/async-load-config';
+import { unlinkSync, writeFileSync } from 'fs';
 function cli(args, cwd): Promise<{ code: number; error: ExecException; stdout: string; stderr: string }> {
 	return new Promise((resolve) => {
 		exec(
