+++ src/cli/utils/async-load-config.ts
@@ -1,12 +1,15 @@
 import { ConfigModule } from 'core/types';
 import { getConfigFile } from 'medusa-core-utils';
+import { MonitoringOptions } from 'modules/monitoring';
 
-export async function asyncLoadConfig(): Promise<ConfigModule> {
+type MonitoredConfigModule = ConfigModule & { monitoring: MonitoringOptions };
+
+export async function asyncLoadConfig(): Promise<MonitoredConfigModule> {
 	const configuration = getConfigFile(process.cwd(), `medusa-config`) as {
 		configModule: ConfigModule;
 		configFilePath: string;
 	};
-	const resolveConfigProperties = async (obj: any): Promise<ConfigModule> => {
+	const resolveConfigProperties = async (obj: any): Promise<MonitoredConfigModule> => {
 		for (const key of Object.keys(obj)) {
 			if (typeof obj[key] === 'object' && obj[key] !== null) {
 				await resolveConfigProperties(obj[key]);
