+++ src/cli/utils/tests/async-load-config.spec.ts
@@ -1,14 +1,8 @@
-import { unlinkSync, writeFileSync } from 'fs';
+import { readFileSync, unlinkSync, writeFileSync } from 'fs';
 import { asyncLoadConfig } from '../async-load-config';
-import {
-	asyncDataWithAsyncFunctionAndAsyncParameter,
-	asyncFunctionWithAsyncData,
-	asyncyFunctionWithNonAsyncData,
-} from '../fixtures/asyn-load-config-test-fixture';
-const pathToConfigFile = `${process.cwd()}/medusa-config.js`;
-const runLoadTest = async (testFixture: string) => {
-	const dummyMedusaConfig = testFixture;
 
+const pathToConfigFile = `${process.cwd()}/medusa-config.js`;
+const runLoadTest = async (testFixtureFileName: string) => {
 	const testProjectConfig = {
 		database_database: './medusa-db.sql',
 		database_type: 'sqlite',
@@ -16,8 +10,8 @@ const runLoadTest = async (testFixture: string) => {
 		store_cors: 'STORE_CORS',
 		admin_cors: 'ADMIN_CORS',
 	};
-
-	writeFileSync(pathToConfigFile, dummyMedusaConfig);
+	const fixtureData = readFileSync(`${__dirname}/../fixtures/${testFixtureFileName}.js`);
+	writeFileSync(pathToConfigFile, fixtureData);
 	const fixture = require.resolve(pathToConfigFile);
 	expect(fixture).toBeDefined();
 	const configModule = await asyncLoadConfig();
@@ -27,6 +21,7 @@ const runLoadTest = async (testFixture: string) => {
 	expect(configModule.projectConfig.database_password).not.toBeInstanceOf(Promise);
 	expect(configModule.projectConfig.database_password.length).toBeGreaterThan(0);
 	expect(configModule.projectConfig).toMatchObject(testProjectConfig);
+	unlinkSync(pathToConfigFile);
 };
 
 describe('async load tests', () => {
@@ -35,18 +30,14 @@ describe('async load tests', () => {
 		jest.clearAllTimers();
 	});
 
-	afterEach(() => {
-		unlinkSync(pathToConfigFile);
-	});
-
 	it('should aysnc load medusa-config with non-async', async () => {
-		await runLoadTest(asyncyFunctionWithNonAsyncData);
+		await runLoadTest('asyncyFunctionWithNonAsyncData');
 	});
 
 	it('should aysnc load medusa-config with async data', async () => {
-		await runLoadTest(asyncFunctionWithAsyncData);
+		await runLoadTest('asyncFunctionWithAsyncData');
 	});
 	it('should aysnc load medusa-config with async data with async function', async () => {
-		await runLoadTest(asyncDataWithAsyncFunctionAndAsyncParameter);
+		await runLoadTest('asyncDataWithAsyncFunctionAndAsyncParameter');
 	});
 });
